requirements.txt
Django==3.2.3
django-filter==23.1
djangorestframework==3.12.4
PyJWT==2.1.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
